requirements.txt
Django==1.9.6
MySQL-python==1.2.5
Twisted==16.2.0
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
asgi-redis==0.13.1
asgiref==0.13.3
autobahn==0.14.1
billiard==3.3.0.23
celery==3.1.23
channels==0.14.2
daphne==0.12.1
django-braces==1.9.0
django-cleanup==0.4.2
django-crispy-forms==1.6.0
django-datatables-view==1.13.0
django-debug-toolbar==1.4
django-guardian==1.4.4
django-mail-templated==2.6.2
djangorestframework==3.3.3
kombu==3.0.35
msgpack-python==0.4.7
pytz==2016.4
redis==2.10.5
six==1.10.0
sqlparse==0.1.19
txaio==2.5.1
wsgiref==0.1.2
zope.interface==4.2.0
